+++ src/array/pluck.js
@@ -1,10 +1,10 @@
-define(['./map', '../function/prop'], function (map, prop) {
+define(['./map'], function (map) {
 
     /**
      * Extract a list of property values.
      */
     function pluck(arr, propName){
-        return map(arr, prop(propName));
+        return map(arr, propName);
     }
 
     return pluck;