+++ src/function.js
@@ -7,6 +7,7 @@ return {
     'compose' : require('./function/compose'),
     'debounce' : require('./function/debounce'),
     'func' : require('./function/func'),
+    'makeIterator_' : require('./function/makeIterator_'),
     'partial' : require('./function/partial'),
     'prop' : require('./function/prop'),
     'series' : require('./function/series'),