+++ tests/spec/spec-function.js
@@ -5,6 +5,7 @@ define([
     'function/spec-compose',
     'function/spec-debounce',
     'function/spec-func',
+    'function/spec-makeIterator_',
     'function/spec-partial',
     'function/spec-prop',
     'function/spec-series',